simplify_repository_1
adapter > outbound > repository 분기 처리 및 코드 정리
브랜치 변경안함
