+++ adapters/outbound/index.js
@@ -1,6 +1,7 @@
 // outbound 클래스 전달 객체
+
 module.exports = {
-    Auth: require('./auth'),
-    Repository: require('./repository'),
+    auth: require('./auth'),
+    repo: require('./repository'),
     // SendMail: require('./sendMail'),
 };
